make bot more modular
